+++ src/main/java/com/nhnacademy/illuwa/d_book/book/entity/Book.java
@@ -6,6 +6,7 @@ import com.nhnacademy.illuwa.d_book.tag.entity.BookTag;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -41,10 +42,10 @@ public class Book {
     private String isbn;
 
     @Column(nullable = false)
-    private int regularPrice;
+    private BigDecimal regularPrice;
 
     @Column(nullable = false)
-    private int salePrice;
+    private BigDecimal salePrice;
 
     @OneToMany(mappedBy = "book", cascade = CascadeType.PERSIST)
     private List<BookImage> bookImages;
